NHP quant, yeast SA quant graphs/analysis, new qPCR graphs
Decided to finally bit the bullet and build a pipeline for SA quantification from a plate layout. Used the preformed biofilm script as a base. The dose-dependent response of Acinetobacter SA_lux loses effectiveness as the concentration of SA in the well increases. So I spent so long learning non-linear regressions to get a good fit for the SA biosensor curves. But in the end the best fit was linear so all hard work was for nothing... always do the easy thing first.